+++ src/solvers/gmres.jl
@@ -1,4 +1,5 @@
 import Krylov
+import Suppressor
 
 struct GMRES{T} <: LinearMap{T}
     A
@@ -15,6 +16,14 @@ end
 
 Base.axes(solver::GMRES) = reverse(axes(solver.A))
 Base.size(solver::GMRES) = reverse(size(solver.A))
+
+function LinearAlgebra.mul!(y::AbstractVecOrMat, solver::GMRES, x::AbstractVector)
+    fill!(y,0)
+    y, ch = solve!(y, solver, x)
+    solver.verbose > 0 && println("Number of iterations: ", ch.niter)
+    ch.solved || error("Iterative solver did not converge.")
+    return y
+end
 
 
 function GMRES(A;
@@ -35,18 +44,28 @@ end
 
 
 function solve!(x, solver::GMRES, b)
-    Krylov.gmres!(solver.workspace, solver.A, Vector(b);
-        M = solver.M, 
-        N = solver.N,
-        # memory = solver.memory, 
-        restart = solver.restart, 
-        atol = solver.atol, 
-        rtol = solver.rtol, 
-        itmax = solver.itmax, 
-        verbose = solver.verbose,
-        history = true)
-    x .= Krylov.solution(solver.workspace)
-    return x, Krylov.statistics(solver.workspace)
+
+    m, n = size(solver.A)
+    T = promote_type(eltype(solver.A), eltype(b))
+    # @show T
+    ws = Krylov.GmresWorkspace(m, n, Vector{T}; solver.memory)
+    # @show typeof(ws)
+
+    Suppressor.@suppress_err begin
+        Krylov.gmres!(ws, solver.A, Vector(b);
+            M = solver.M, 
+            N = solver.N,
+            # memory = solver.memory, 
+            restart = solver.restart, 
+            atol = solver.atol, 
+            rtol = solver.rtol, 
+            itmax = solver.itmax, 
+            verbose = solver.verbose,
+            history = true)
+    end
+    x .= Krylov.solution(ws)
+
+    return x, Krylov.statistics(ws)
 end
 
 function solve(solver::GMRES, b)
@@ -61,7 +80,7 @@ end
 
 @testitem "GMRES" begin
     using LinearAlgebra
-    
+
     A = rand(10, 10)
     b = rand(10)